# Separate driver (hardware comms) from buffer (memory mgmt) and enable double buffering on MicroGraphics