--- a/Source/Meadow.Foundation.Core/Displays/IPixelBuffer.cs
+++ b/Source/Meadow.Foundation.Core/Displays/IPixelBuffer.cs
@@ -6,13 +6,19 @@ using System.Text;
 namespace Meadow.Foundation.Graphics.Buffers
 {
     /// <summary>
-    /// An interface representing the concept of a
-    /// buffer that holds a collection of pixels
-    /// that can be rendered to a wide variety of devices
-    /// at a variety of bit depths.
+    /// IPixelBuffer provides a standard interface for
+    /// representing the display state of a device capable of
+    /// displaying pixels. It specifies methods for performing
+    /// common primative operations on a buffer of pixel data.
     /// 
-    /// This interface has default implementations for many
-    /// of its properties and methods.
+    /// Conceptually, implementing classes should:
+    /// 
+    /// 1. Specify a bit depth for pixels
+    /// 2. Specify a color mode
+    /// 3. Preserve the display state as a byte[] in the PixelBuffer
+    /// 4. Optimize primative drawing methods for the bit depth of pixels
+    /// 5. Be abstracted/decoupled from a specific device driver
+    /// 
     /// </summary>
     public interface IPixelBuffer
     {