--- a/Source/Meadow.Foundation.Libraries_and_Frameworks/Graphics/MicroGraphics/Driver/Graphics.MicroGraphics/Buffers/Buffer1bpp.cs
+++ b/Source/Meadow.Foundation.Libraries_and_Frameworks/Graphics/MicroGraphics/Driver/Graphics.MicroGraphics/Buffers/Buffer1bpp.cs
@@ -1,12 +1,12 @@
-﻿using System;
+﻿using Meadow.Foundation.Displays;
+using MicroGraphics.Buffers;
+using System;
 
 namespace Meadow.Foundation.Graphics.Buffers
 {
-    public class Buffer1bpp : BufferBase
+    public class Buffer1bpp : PixelBufferBase
     {
-        public override int ByteCount => Width * Height / 8;
-
-        public override ColorType displayColorMode => ColorType.Format1bpp;
+        public override ColorType ColorMode => ColorType.Format1bpp;
 
         public Buffer1bpp(int width, int height, byte[] buffer) : base(width, height, buffer) { }
 
@@ -23,11 +23,53 @@ namespace Meadow.Foundation.Graphics.Buffers
             Buffer = new byte[bufferSize];
         }
 
-        public bool GetPixelIsColored(int x, int y)
+        public void Clear(bool isColored)
         {
-            var index = (y >> 8) * Width + x;
+            // split the color in to two byte values
+            Buffer[0] = (byte)(isColored ? 0xFF : 0);
 
-            return (Buffer[index] & (1 << y % 8)) != 0;
+            int arrayMidPoint = Buffer.Length / 2;
+            int copyLength;
+
+            for (copyLength = 1; copyLength < arrayMidPoint; copyLength <<= 1)
+            {
+                Array.Copy(Buffer, 0, Buffer, copyLength, copyLength);
+            }
+
+            Array.Copy(Buffer, 0, Buffer, copyLength, Buffer.Length - copyLength);
+        }
+
+        public override void Fill(Color color)
+        {
+            Clear(color.Color1bpp);
+        }
+
+        public override void Fill(Color color, int x, int y, int width, int height)
+        {
+            if (x < 0 || x + width > Width ||
+                y < 0 || y + height > Height)
+            {
+                throw new ArgumentOutOfRangeException();
+            }
+
+            var isColored = color.Color1bpp;
+
+            for (int i = 0; i < width; i++)
+            {
+                for (int j = 0; j < height; j++)
+                {   //byte aligned and at least 8 rows to go
+                    if ((j + y) % 8 == 0 && j + y + 8 <= height)
+                    {
+                        //set an entire byte - fast
+                        Buffer[((j + y) >> 3) * Width + x + i] = (byte)(isColored ? 0xFF : 0);
+                        j += 7; //the main loop will add 1 to make it 8
+                    }
+                    else
+                    {
+                        SetPixel(x + i, y + j, isColored);
+                    }
+                }
+            }
         }
 
         public override Color GetPixel(int x, int y)
@@ -54,80 +96,48 @@ namespace Meadow.Foundation.Graphics.Buffers
             SetPixel(x, y, color.Color1bpp);
         }
 
-        public override void Fill(Color color)
+        public override void InvertPixel(int x, int y)
         {
-            Clear(color.Color1bpp);
+            throw new NotImplementedException();
         }
 
-        public override void Fill(Color color, int x, int y, int width, int height)
+        public new void WriteBuffer(int x, int y, IPixelBuffer buffer)
         {
-            if (x < 0 || x + width > Width ||
-                y < 0 || y + height > Height)
+            // our color modes match, we can fast copy
+            if(buffer.ColorMode == ColorMode)
             {
-                throw new ArgumentOutOfRangeException();
-            }
-
-            var isColored = color.Color1bpp;
-            
-            for (int i = 0; i < width; i++)
-            {
-                for (int j = 0; j < height; j++)
-                {   //byte aligned and at least 8 rows to go
-                    if((j + y) % 8 == 0 && j + y + 8 <= height)
-                    {
-                        //set an entire byte - fast
-                        Buffer[((j + y) >> 3) * Width + x + i] = (byte)(isColored ? 0xFF : 0);
-                        j += 7; //the main loop will add 1 to make it 8
-                    }
-                    else
-                    {
-                        SetPixel(x + i, y + j, isColored);
-                    }
-                }
-            }
-        }
-
-        public void Clear(bool isColored)
-        {
-            // split the color in to two byte values
-            Buffer[0] = (byte)(isColored ? 0xFF : 0);
-
-            int arrayMidPoint = Buffer.Length / 2;
-            int copyLength;
-
-            for (copyLength = 1; copyLength < arrayMidPoint; copyLength <<= 1)
-            {
-                Array.Copy(Buffer, 0, Buffer, copyLength, copyLength);
-            }
-
-            Array.Copy(Buffer, 0, Buffer, copyLength, Buffer.Length - copyLength);
-        }
-
-        public new void WriteBuffer(int x, int y, IDisplayBuffer buffer)
-        {
-            if (base.WriteBuffer(x, y, buffer))
-            {   //call the base for validation
-                //and to handle the slow path when buffers don't match
-                return;
-            }
-
-            for (int i = 0; i < buffer.Width; i++)
-            {
-                for (int j = 0; j < buffer.Height; j++)
+                for (int i = 0; i < buffer.Width; i++)
                 {
-                    //if we got really clever we could find other alignment points but this is a good start
-                    if(y%8 == 0 && j + 8 <= buffer.Height)
+                    for (int j = 0; j < buffer.Height; j++)
                     {
-                        //copy an entire byte - fast
-                        Buffer[((y + j) >> 3) * Width + x + i] = buffer.Buffer[(j >> 3) * buffer.Width + i];
-                        j += 7; //the main loop will add 1 to make it 8
-                    }
-                    else
-                    {   //else 1 bit at a time 
-                        SetPixel(x + i, y + j, (buffer as Buffer1bpp).GetPixelIsColored(i, j));
+                        //if we got really clever we could find other alignment points but this is a good start
+                        if (y % 8 == 0 && j + 8 <= buffer.Height)
+                        {
+                            //copy an entire byte - fast
+                            Buffer[((y + j) >> 3) * Width + x + i] = buffer.Buffer[(j >> 3) * buffer.Width + i];
+                            j += 7; //the main loop will add 1 to make it 8
+                        }
+                        else
+                        {   //else 1 bit at a time 
+                            SetPixel(x + i, y + j, (buffer as Buffer1bpp).GetPixelIsColored(i, j));
+                        }
                     }
                 }
             }
+            // no match, fall through to default slow copy
+            else
+            {
+                base.WriteBuffer(x, y, buffer);
+            }
+        }
+
+
+
+        public bool GetPixelIsColored(int x, int y)
+        {
+            var index = (y >> 8) * Width + x;
+
+            return (Buffer[index] & (1 << y % 8)) != 0;
         }
 
         public Buffer1bpp Rotate(RotationType rotation)