--- a/Source/Meadow.Foundation.Libraries_and_Frameworks/Graphics/MicroGraphics/Driver/Graphics.MicroGraphics/Buffers/BufferGray4.cs
+++ b/Source/Meadow.Foundation.Libraries_and_Frameworks/Graphics/MicroGraphics/Driver/Graphics.MicroGraphics/Buffers/BufferGray4.cs
@@ -1,6 +1,4 @@
-﻿using Meadow.Foundation.Displays;
-using MicroGraphics.Buffers;
-using System;
+﻿using System;
 
 namespace Meadow.Foundation.Graphics.Buffers
 {