--- a/Source/Meadow.Foundation.Libraries_and_Frameworks/Graphics/MicroGraphics/Driver/Graphics.MicroGraphics/Buffers/BufferRgb444.cs
+++ b/Source/Meadow.Foundation.Libraries_and_Frameworks/Graphics/MicroGraphics/Driver/Graphics.MicroGraphics/Buffers/BufferRgb444.cs
@@ -106,7 +106,33 @@ namespace Meadow.Foundation.Graphics.Buffers
 
         public override void InvertPixel(int x, int y)
         {
-            throw new NotImplementedException();
+            byte r, g, b;
+            int index;
+            var pBuffer = Buffer;
+            if (x % 2 == 0)
+            {
+                index = (x + y * Width) * 3 / 2;
+
+                r = (byte)(pBuffer[index] >> 4);
+                g = (byte)(pBuffer[index] & 0x0F);
+                b = (byte)(pBuffer[index + 1] >> 4);
+            }
+            else
+            {
+                index = ((x - 1 + y * Width) * 3 / 2) + 1;
+                r = (byte)(pBuffer[index] & 0x0F);
+                g = (byte)(pBuffer[index + 1] >> 4);
+                b = (byte)(pBuffer[index + 1] & 0x0F);
+            }
+
+            r = (byte)(~r & 0x0F);
+            g = (byte)(~g & 0x0F);
+            b = (byte)(~b & 0x0F);
+
+            //get new color
+            var color = (ushort)(r << 8 | g << 4 | b);
+
+            SetPixel(x, y, color);
         }
 
         public override void WriteBuffer(int x, int y, IPixelBuffer buffer)