--- a/Source/Meadow.Foundation.Libraries_and_Frameworks/Graphics/MicroGraphics/Driver/Graphics.MicroGraphics/Buffers/PixelBufferBase.cs
+++ b/Source/Meadow.Foundation.Libraries_and_Frameworks/Graphics/MicroGraphics/Driver/Graphics.MicroGraphics/Buffers/PixelBufferBase.cs
@@ -1,7 +1,4 @@
-﻿using Meadow.Foundation;
-using Meadow.Foundation.Displays;
-using Meadow.Foundation.Graphics;
-using System;
+﻿using System;
 
 namespace Meadow.Foundation.Graphics.Buffers
 {