fixes https://github.com/WildernessLabs/Meadow.Foundation/issues/338
fixes https://github.com/WildernessLabs/Meadow.Foundation/issues/333
fixes https://github.com/WildernessLabs/Meadow.Foundation/issues/336

This PR _mostly_ fixes the issues above. However, there were some differences between concept and where the rubber meets the road! I would have loved to do this in more granular PRs but I couldn't figure out a way to have smaller PRs without broken builds or general confusion.

### Changes

- The `IDisplayBuffer` interface is now called `IPixelBuffer`
    - Added new properties
    - Added implementation suggestions as a docs comment
    - All references in the main `MeadowGraphics.sln` across 291 projects have been updated
    - All buffers have been updated to satisfy the interface
    - The `BufferBase` class was renamed to `PixelBufferBase` and has more helper properties to provide things like BitDepth and ByteCount
    - All buffer implementations no longer do method hiding via the `new` keyword on the `WriteBuffer` method. This had unpredictable behavior and was hard to debug.
- The `IGraphicsDisplay` interface is now called `IGraphicsDriver`
    - Added new properties
    - Added implementation suggestions as a docs comment
    - All references in the main `MeadowGraphics.sln` across 291 projects have been updated
    - All drivers have been updated to satisfy the `IGraphicsDriver` interface.
    - The interface has an `InvertPixel` method, most of the buffers throw a NotImplementedException for now
    - Graphics driver pixel manipulation methods attempt to just wrap/call into the buffer methods where possible. However, I avoided touching drivers more than minor tweaks to satisfy the interface. Many drivers don't even use an `IPixelBuffer`.
- `TftSpiBase` has numerous improvements that moves depth-specific buffer manipulation into the appropriate buffer instead of hardcoded manipulations at the driver level
- MicroGraphics now supports double-buffering
    - MG has an `IGraphicsDriver driver` that represents the hardware device
    - MG has an `IPixelBufffer pixelBuffer` that represents the draw buffer
    - By default `pixelBuffer == driver.PixelBuffer` so the external API is identical
    - When `Show` is called, the `pixelBuffer` will be copied to the `driver.PixelBuffer` if they are not the same
    - `driver` and `pixelBuffer` properties are now protected instead of private
    - Extending classes can now do advanced double and multi-buffering

### Tests

- I wrote a [MicroGraphics test app as part of Glade2d](https://github.com/profexorgeek/Glade2d/blob/master/Glade2d/MicroGraphicsTest.cs) and ran it before and after on my device (7789)
    - Before the refactor 444 color mode averaged ~950ms to complete the draw test
    - After the refactor 444 color mode averaged ~850ms to complete the draw test
    - Before the refactor 565 color mode averaged ~350ms to complete the draw test
    - After the refactor 565 color mode averaged ~250ms to complete the draw test
    - I attribute speed improvements to a simplified code path with less method chaining and some minor perf optimizations in both the spi base class and the 444 and 565 buffers
- I ran unit tests, the same number were broken before and after
- I ran the whole Glade2d renderer against it, no significant FPS change but Glade2d does a lot of it's own drawing so this is not a surprise

### NOT Fixed

- My original goal was to remove primitive draw ops from drivers entirely. In practice, this was too big of a change and would make simple LED displays more complex. I tried to satisfy interfaces without touching those heavily since I lack the devices to test. If devices want to satisfy the conceptual split better, this can happen over time.
- External projects may need to make minor API changes to reference and satisfy the new interfaces. Method names should be pretty easy to discover and resolve.
- My device only supports 444 and 565 color mode. I mostly reorganized buffers but there were some overlapping methods that I made choices to combine. **I almost definitely broke things here that I can't test and catch!**
- 3/6 unit tests are failing in MicroGraphics. These are failing on main so I don't think I broke them and I didn't attempt to fix them in this PR


UPDATE: Added a late commit with some forgotten changes. These last adds make the MicroGraphics class more extensible.